@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 extension StateExt on State {
   void doPostFrame(Function func) {
-    WidgetsBinding.instance?.addPostFrameCallback((timeStamp) {
+    WidgetsBinding.instance.addPostFrameCallback((timeStamp) {
       func();
     });
   }
